# Faster evaluation of B-splines