--- a/src/BSplines/evaluate_all.jl
+++ b/src/BSplines/evaluate_all.jl
@@ -21,10 +21,10 @@ More precisely:
 
 - `i` is the index of the first B-spline knot ``t_{i}`` when going from ``x``
   towards the left.
-  In other words, it is such that ``t_{i} ≤ x < t{i + 1}``.
+  In other words, it is such that ``t_{i} ≤ x < t_{i + 1}``.
 
   It is effectively computed as `i = searchsortedlast(knots(B), x)`.
-  If the right value of `i` is already known, one can avoid this computation by
+  If the correct value of `i` is already known, one can avoid this computation by
   manually passing this index via the optional `ileft` keyword argument.
 
 - `bs` is a tuple of B-splines evaluated at ``x``:
@@ -42,6 +42,10 @@ More precisely:
 One can pass the optional `op` argument to compute B-spline derivatives instead
 of the actual B-spline values.
 
+## Examples
+
+See [`AbstractBSplineBasis`](@ref) for some examples using the alternative
+evaluation syntax `B(x, [op], [T]; [ileft])`, which calls this function.
 """
 @propagate_inbounds function evaluate_all(
         B::BSplineBasis, x::Real, op::Derivative, ::Type{T}; kws...,