… docs/src/bsplines.md
… src/BandedTensors/BandedTensors.jl
… src/Collocation/Collocation.jl
… src/Galerkin/projection.jl
… src/Galerkin/quadratic.jl
… src/Galerkin/quadratures.jl
… src/Recombinations/bases.jl
… src/Recombinations/matrices.jl
… src/SplineApproximations/SplineApproximations.jl
… src/SplineInterpolations/SplineInterpolations.jl
… src/Splines/spline.jl
… test/airy.jl
… test/bsplines.jl
… test/galerkin.jl
… test/runtests.jl