This PR provides the `evaluate_all` function, which evaluates, at a given point `x`, all B-splines which are non-zero at that point. The function also works for recombined B-splines and for B-spline derivatives. This is much more efficient than evaluating B-splines one by one.

This new function is now used to construct collocation matrices (used in particular for interpolations) and Galerkin tensors.

For convenience, one can also evaluate a B-spline basis `B` using the `B(x)` syntax, which is equivalent to `evaluate_all`.

```julia
using BSplineKit
using BenchmarkTools
using Random

# Construct B-spline basis
rng = MersenneTwister(42)
ξs = sort!(rand(rng, 20))
ξs[begin] = 0; ξs[end] = 1;
B = BSplineBasis(BSplineOrder(6), copy(ξs))

# Evaluate non-zero B-splines at a point
# Note that the number of non-zero B-splines is equal to the B-spline order (k = 6 here)
x = 0.43
evaluate_all(B, x)
# = (14, (2.30799526056475e-6, 0.02175822604119263, 0.2704859488098379, 0.5541008832923826, 0.15224719511338874, 0.0014054387479374313))
```

The first returned element (here `i = 14`) is the index of the *last* B-spline that is non-zero at x.
The second element is a tuple `(b_{i}, b_{i - 1}, …, b{i - k + 1})` with the actual B-spline values.
Note that the values are returned in backwards order!

This syntax is exactly equivalent:

```julia
julia> B(x)
(14, (2.30799526056475e-6, 0.02175822604119263, 0.2704859488098379, 0.5541008832923826, 0.15224719511338874, 0.0014054387479374313))
```

To evaluate derivatives (e.g. the second derivative):

```julia
julia> evaluate_all(B, x, Derivative(2))
(14, (0.9303769749908197, 196.3554353920673, 101.55718315717507, -706.6327432445232, 359.10806040145405, 48.68168731883596))

julia> B(x, Derivative(2))
(14, (0.9303769749908197, 196.3554353920673, 101.55718315717507, -706.6327432445232, 359.10806040145405, 48.68168731883596))
```

One can also use recombined bases:
```julia
julia> R = RecombinedBSplineBasis(Natural(), B)
20-element RecombinedBSplineBasis of order 6, domain [0.0, 1.0]
 knots: [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.160006, 0.172933, 0.176909, 0.262809  …  0.745181, 0.937466, 0.956916, 0.958926, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0]
 BCs left:  (D{2}, D{3})
 BCs right: (D{2}, D{3})

julia> R(x)
(12, (2.30799526056475e-6, 0.02175822604119263, 0.2704859488098379, 0.5541008832923826, 0.15224719511338874, 0.0014054387479374313))
```

Some benchmarks (before/after comparison):

```julia
julia> xs = collocation_points(B);

julia> C = collocation_matrix(B, xs);

julia> @btime collocation_matrix!($C, $B, $xs);
  18.274 μs (0 allocations: 0 bytes)  # before
  1.927 μs (0 allocations: 0 bytes)   # with this PR

julia> G = galerkin_matrix(B);

julia> @btime galerkin_matrix!($G, $B);
  68.334 μs (0 allocations: 0 bytes)  # before
  13.544 μs (0 allocations: 0 bytes)  # with this PR

julia> T = galerkin_tensor(B, Derivative.((0, 0, 1)));

julia> @btime galerkin_tensor!($T, $(B, B, B), Derivative.((0, 0, 1)));
  225.285 μs (272 allocations: 9.73 KiB)  # before
  109.311 μs (1 allocation: 1008 bytes)   # with this PR
```